+++ recommender-back/src/tests/apis/helpers_test.py
@@ -1,5 +1,5 @@
 import unittest
-from apis.helpers import PointOfInterest
+from apis.helpers import Recommender
 
 
 class PointOfInterestTest(unittest.TestCase):
@@ -12,7 +12,7 @@ class PointOfInterestTest(unittest.TestCase):
             'name': 'POI 1',
             'weather': {'Air temperature': '30 °C', 'Humidity': '50%'}
         }
-        poi = PointOfInterest(self.time,**poi_data)
+        poi = Recommender(self.time,**poi_data)
         self.assertAlmostEqual(poi.score, 1.0)
 
     def test_calculate_score_with_valid_data_nothot(self):
@@ -21,7 +21,7 @@ class PointOfInterestTest(unittest.TestCase):
             'name': 'POI 1',
             'weather': {'Air temperature': '10 °C', 'Humidity': '10%'}
         }
-        poi = PointOfInterest(self.time,**poi_data)
+        poi = Recommender(self.time,**poi_data)
         self.assertAlmostEqual(poi.score, 0.0)
 
     def test_calculate_score_with_invalid_data(self):
@@ -30,7 +30,7 @@ class PointOfInterestTest(unittest.TestCase):
             'name': 'POI 2',
             'weather': {'Air temperature': None, 'Humidity': '50%'}
         }
-        poi = PointOfInterest(self.time,**poi_data)
+        poi = Recommender(self.time,**poi_data)
         self.assertEqual(poi.score, -float('inf'))
 
     def test_calculate_score_with_invalid_time(self):
@@ -40,7 +40,7 @@ class PointOfInterestTest(unittest.TestCase):
             'name': 'POI 1',
             'weather': {'Air temperature': '30 °C', 'Humidity': '50%'}
         }
-        poi = PointOfInterest(self.time,**poi_data)
+        poi = Recommender(self.time,**poi_data)
         self.assertAlmostEqual(poi.score, 0.0)
 
     def test_point_of_interest_initialization(self):
@@ -49,7 +49,7 @@ class PointOfInterestTest(unittest.TestCase):
             'name': 'POI 3',
             'weather': {'Air temperature': '25 °C', 'Humidity': '45%'}
         }
-        poi = PointOfInterest(self.time,**poi_data)
+        poi = Recommender(self.time,**poi_data)
         self.assertEqual(poi.id, 3)
         self.assertEqual(poi.name, 'POI 3')
         self.assertEqual(poi.score, 1.0)
